[feat]EditorGNB 분리
